+++ frontend/src/Game/types/match.model.ts
@@ -1,0 +1,1 @@
+export type UPlayer = 'left' | 'right'
